Time: 0 ms (100%), Space: 14.7 MB (30.85%) - LeetHub

+++ 0001-two-sum/0001-two-sum.cpp
@@ -2,9 +2,9 @@ class Solution {
 public:
     vector<int> twoSum(vector<int>& nums, int target) {
         unordered_map<int, int> umap;
-        long long int n = nums.size();
-        for(int i = 0; i < n; i++) {
-            if(umap.find(target - nums[i]) != umap.end()) {
+        long int n = nums.size();
+        for (int i = 0; i < n; i++) {
+            if (umap.find(target - nums[i]) != umap.end()) {
                 return {umap[target - nums[i]], i};
             }
             umap[nums[i]] = i;
